Add git commit template Prior to this change, commits did not follow any predictable pattern
This change adds a .git-commit-message file that provides a default
template to use for commit messages.

The template can be applied by using the following command:
    git config commit.template .git-commit-message

There currently doesn't seem to be support for message templates ig
VS Code when using the source control sidebar, use `git commit` in
console for now.