Add a custom scrubber for html_sanitize_ex
Prior to this change, the default markdown_html scrubber was used.
That scrubber did not allow for some required tags

This change adds a custom scrubber that mostly matchess the default
markdown_html one from html_sanitize_ex but adds support for cites in
blockquotes and address tags as well as some other minor changes.