Add configurable Uploads directory
Prior to this change, the upload directory was hardcoded, in multiple
locations in the application, causing issues with releases and generally
being bad design.

This change improves the situation by making the upload directory
configurable. For production releases, a runtime configuration is
assumed.

It also moves the video out of source control and into the uploads
directory